kindergarten_project/requirements.txt
backcall==0.1.0
decorator==4.3.0
Django==1.11
django-extensions==2.1.0
djangorestframework==3.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
rest-framework-generic-relations==1.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
